Safari-611.1.10/LayoutTests/svg/animations/svgpointlist-animation-1-expected.txt

SVG 1.1 dynamic animation tests

Tests from-to animation of points on polygons.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS poly.animatedPoints.getItem(2).x is 200
PASS poly.animatedPoints.getItem(2).y is 200
PASS poly.points.getItem(2).x is 200
PASS poly.points.getItem(2).y is 200
PASS poly.animatedPoints.getItem(2).x is 150
PASS poly.animatedPoints.getItem(2).y is 150
PASS poly.points.getItem(2).x is 200
PASS poly.points.getItem(2).y is 200
PASS poly.animatedPoints.getItem(2).x is 100
PASS poly.animatedPoints.getItem(2).y is 100
PASS poly.points.getItem(2).x is 200
PASS poly.points.getItem(2).y is 200
PASS poly.animatedPoints.getItem(2).x is 200
PASS poly.animatedPoints.getItem(2).y is 200
PASS poly.points.getItem(2).x is 200
PASS poly.points.getItem(2).y is 200
PASS successfullyParsed is true

TEST COMPLETE
